+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Backward Compatibility Breaks
 ### Added
 * Added support for numeric `minimum_should_match` values in `TermsSet` query [#2293](https://github.com/ruflin/Elastica/pull/2293)
+* Added support for "search after" based pagination [#1645](https://github.com/ruflin/Elastica/issues/1645)
 ### Changed
 ### Deprecated
 ### Removed